+++ packages/kb.gbapp/dialogs/AskDialog.ts
@@ -262,7 +262,7 @@ export class AskDialog extends IGBDialog {
 
         // Tries to answer by NLP.
 
-        let nextDialog = await min.conversationalService['routeNLP2'](step, min, text);
+        let nextDialog = await min.conversationalService.routeNLP(step, min, text);
         if (nextDialog) {
           return nextDialog;
         }